The Kata is done.

+++ ParkingLotKata2/ParkingLotKata2/Driver.cs
@@ -9,7 +9,7 @@ namespace ParkingLotKata2
             Wallet += money;
         }
 
-        public void Withdraw(int money)
+        public void Withdraw(double money)
         {
             Wallet -= money;
         }
