+++ ParkingLotKata2/ParkingLotKata2/IDriver.cs
@@ -2,6 +2,6 @@ namespace ParkingLotKata2
 {
     public interface IDriver
     {
-        void Withdraw(int money);
+        void Withdraw(double money);
     }
 }
