+++ ParkingLotKata2/ParkingLotKata2/IVehicle.cs
@@ -3,6 +3,6 @@ namespace ParkingLotKata2
     public interface IVehicle
     {
         IDriver Driver { get; }
-        int Length { get;  }
+        double Length { get; }
     }
 }
